add bias

+++ include/ck/tensor_operation/gpu/element/binary_element_wise_operation.hpp
@@ -33,7 +33,7 @@ struct Add
     __host__ __device__ constexpr void
     operator()<float>(float& y, const float& x0, const half_t& x1) const
     {
-        y = x0 + type_convert<half_t>(x1);
+        y = x0 + type_convert<float>(x1);
     };
 
     template <>
